# api/openapi.yaml
openapi: 3.0.3
info:
  title: iRODS REST API
  version: 0.1.0
  description: |
    REST-ful iRODS running on GoLang.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/server:
    get:
      operationId: getServerInfo
      tags: [health]
      summary: Fetch iRODS server information
      description: |
        Returns iRODS server version details (similar to `miscsvrinfo`) plus
        service-side iRODS connection configuration details.
      security:
        - BearerAuth: []
        - BasicAuth: []
      responses:
        "200":
          description: Server information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerInfoResponse"
        "403":
          description: Access denied while reading server information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/resource:
    get:
      operationId: getResources
      tags: [path]
      summary: List resource servers in the configured iRODS zone
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: scope
          required: false
          schema:
            type: string
            enum: [top, all]
            default: top
          description: |
            Resource listing scope. `top` is the default view intended for top-level
            resources. `all` returns the full visible flat resource set. The current
            backend resource model is flat, so these values are currently equivalent.
      responses:
        "200":
          description: Resource servers visible in the zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceCollectionResponse"
        "403":
          description: Access denied while listing resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/resource/{resource_id}:
    get:
      operationId: getResource
      tags: [path]
      summary: Fetch details for one iRODS resource
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: resource_id
          required: true
          schema:
            type: string
          description: Resource identifier. This currently uses the iRODS resource name.
      responses:
        "200":
          description: Resource details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceResponse"
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while reading the resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/principal:
    get:
      operationId: getPrincipals
      tags: [user]
      summary: Search iRODS users and groups
      description: |
        Searches user and group principals by name prefix for combined selector
        UIs. Use the separate `/api/v1/user` and `/api/v1/usergroup` prefix
        searches when a combined result set is not needed.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: query
          required: true
          schema:
            type: string
            minLength: 3
          description: Principal name prefix to search.
        - in: query
          name: kind
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [user, group]
          style: form
          explode: true
          description: Optional principal kind filter. May be repeated.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum result count.
      responses:
        "200":
          description: Matching principals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrincipalSearchResponse"
        "400":
          description: Invalid search option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while searching principals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user:
    get:
      operationId: getUsers
      tags: [user]
      summary: List iRODS users
      description: |
        Lists visible iRODS user principals in the configured zone. When `prefix`
        is supplied for autocomplete-style search, it must be at least 3 characters.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [rodsuser, rodsadmin]
          description: Optional user type filter.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional username prefix filter for autocomplete.
      responses:
        "200":
          description: Users visible in the zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCollectionResponse"
        "400":
          description: Invalid user list option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postUser
      tags: [user]
      summary: Create one iRODS user
      description: |
        Requires the authenticated iRODS user to have type `rodsadmin` or `groupadmin`.
        Synchronization clients may use `reconcile=true` to make creation repeatable
        when the existing principal is already the requested user type. The operation
        remains a conflict when the existing principal is a group or when the existing
        user type differs from the requested type.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Treat an existing user with the requested type
            as success instead of returning 409. A type mismatch or non-user
            principal with the same name remains a conflict. If a password is supplied
            for an existing matching user, the password is applied.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateRequest"
      responses:
        "200":
          description: Desired user state satisfied when reconciliation mode is used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid create request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin or groupadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user/me:
    get:
      operationId: getCurrentUserMembership
      tags: [user]
      summary: Fetch the authenticated iRODS user and groups
      description: |
        Returns the effective authenticated iRODS user, group memberships, and
        computed administration flags for Starbase user/group UI decisions.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum group count.
      responses:
        "200":
          description: Current iRODS user membership and administration flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserMembershipResponse"
        "400":
          description: Invalid current-user option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while fetching current user membership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user/membership-summary:
    get:
      operationId: getUserMembershipSummaries
      tags: [user]
      summary: List users with group membership summaries
      description: |
        Returns user rows with group names for list-scale administration UIs.
        This avoids broad client-side joins or repeated per-user group lookups.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: type
          required: false
          schema:
            type: string
            enum: [rodsuser, rodsadmin]
          description: Optional user type filter.
        - in: query
          name: prefix
          required: false
          schema:
            type: string
            minLength: 3
          description: Optional username prefix filter.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum user count.
      responses:
        "200":
          description: Users with membership summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserMembershipSummaryCollectionResponse"
        "400":
          description: Invalid summary option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing user memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user/{user_name}:
    get:
      operationId: getUser
      tags: [user]
      summary: Fetch one iRODS user
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while reading the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: putUser
      tags: [user]
      summary: Update one iRODS user
      description: |
        Requires the authenticated iRODS user to have type `rodsadmin`.
        Synchronization clients may use `reconcile=true` as a constrained upsert:
        a missing user is created only when the request includes `type`. Password-only
        updates still require the user to exist, and existing non-user principals or
        incompatible types remain guarded by normal error handling.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Create the user when missing if the request
            includes a target `type`; password-only updates still return 404 when
            the user is absent. Reconciliation does not bypass rodsadmin authorization.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid update request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteUser
      tags: [user]
      summary: Delete one iRODS user
      description: |
        Requires the authenticated iRODS user to have type `rodsadmin` or `groupadmin`.
        Synchronization clients may use `reconcile=true` so deleting an already absent
        user is reported as success. The operation does not delete groups or other
        non-user principals with the same name.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone. Defaults to the configured iRODS zone.
        - in: query
          name: reconcile
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Desired-state sync mode. Treat a missing user, or an existing non-user
            principal with this name, as already absent and return success.
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Authenticated iRODS user is not rodsadmin or groupadmin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/user/{user_name}/usergroup:
    get:
      operationId: getUserGroupsForUser
      tags: [user]
      summary: List groups containing one user
      description: |
        Returns the iRODS groups containing the selected user. This is the
        reverse-membership lookup for user detail and administration UIs.
      security:
        - BearerAuth: []
        - BasicAuth: []
      parameters:
        - in: path
          name: user_name
          required: true
          schema:
            type: string
          description: iRODS username.
        - in: query
          name: zone
          required: false
          schema:
            type: string
          description: iRODS zone to search. Defaults to the configured iRODS zone.
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum group count.
      responses:
        "200":
          description: Groups containing the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserGroupRefCollectionResponse"
        "400":
          description: Invalid reverse membership option
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Access denied while listing user groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      scheme: basic
  schemas:
...
    ServerInfoResponse:
      type: object
      required: [server_info]
      properties:
        server_info:
          $ref: "#/components/schemas/ServerInfo"
...
    ResourceResponse:
      type: object
      required: [resource]
      properties:
        resource:
          $ref: "#/components/schemas/Resource"
    ResourceCollectionResponse:
      type: object
      required: [resources, count, scope, links]
      properties:
        resources:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        count:
          type: integer
        scope:
          type: string
          enum: [top, all]
        links:
          type: object
          properties:
            self:
              $ref: "#/components/schemas/ActionLink"
...
    UserCreateRequest:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
          description: iRODS username to create.
        type:
          type: string
          enum: [rodsuser, rodsadmin]
          description: iRODS user type for the new user.
        password:
          type: string
          format: password
          description: Optional initial iRODS user password.
...
    PrincipalSearchResponse:
      type: object
      required: [principals, count, zone, query]
      properties:
        principals:
          type: array
          items:
            $ref: "#/components/schemas/PrincipalSearchResult"
        count:
          type: integer
        zone:
          type: string
        query:
          type: string
        kinds:
          type: array
          items:
            type: string
            enum: [user, group]
        limit:
          type: integer
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        fields:
          type: object
          description: Optional field-level validation messages keyed by request field name.
          additionalProperties:
            type: string